PointSurfaceReconstruction/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
tqdm>=4.62.0
matplotlib>=3.5.0
open3d>=0.15.0
trimesh>=3.9.0
tensorboard>=2.7.0